safari-534.53-branch/LayoutTests/svg/custom/immutable-properties-expected.txt

Tests whether immutable properties can not be modified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewport.x is 0
PASS viewport.x is 100
PASS svgDoc.documentElement.viewport.x is 0
PASS zoomRectScreen.x is 0
PASS zoomRectScreen.x is 100
PASS zoomEvent.zoomRectScreen.x is 0
PASS zoomEvent.previousScale is 0
PASS zoomEvent.previousScale is 0
PASS previousTranslate.x is 0
PASS previousTranslate.x is 300
PASS zoomEvent.previousTranslate.x is 0
PASS zoomEvent.newScale is 0
PASS zoomEvent.newScale is 0
PASS newTranslate.x is 0
PASS newTranslate.x is 300
PASS zoomEvent.newTranslate.x is 0
PASS successfullyParsed is true

TEST COMPLETE
